requirements.txt
cffi==1.17.1
numpy==2.1.2
opuslib==3.0.1
paho-mqtt==1.6.1
playsound==1.3.0
pycparser==2.22
pydub==0.25.1
pygame==2.6.1
setuptools==75.2.0
sounddevice==0.5.1
websockets==13.1
wheel==0.44.0
